create wallet component + shuffle columns around
